+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- `FirClassLikeSymbol<*>.toFirTypeRef(): FirTypeRef`
 - `FirClassLikeSymbol<*>.toValueParameter(): FirValueParameter`
 - `FirFunctionSymbol<*>.contextReceiversToValueParameters(): List<FirValueParameter>`
 - `ClassId.toConeType(): ConeClassLikeType`